requirements.txt
auto_mix_prep==0.2.0
GitPython==3.1.44
openai==1.65.4
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
tiktoken==0.7.0
